pyproject.toml
[tool.poetry]
name = "hw-04"
version = "0.1.0"
description = ""
authors = ["Лизанька <fedko.liza@mail.ru>"]


[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.16"
sqlalchemy = "^2.0"
asyncpg = "^0.30.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}
unicorn = "^2.1.3"
gunicorn = "^23.0.0"
uvicorn = "^0.34.2"
pydantic-settings = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
